+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Portrait
 * Tile - an alternative to Pod, with a drop shadow.
 * Help - A question mark icon with a tooltip.
+* Tooltip
 
 ## Layout Updates
 
@@ -25,6 +26,11 @@
 
 * Allow Tables to shrink in size using the `shrink` prop.
 * Child components of Row can now use a `columnAlign` prop.
+
+
+## New Decorators
+
+* Tooltip Decorator - currently available on Icon and Textbox.
 
 ## Bug Fixes